Run ChannelCrashTests on Linux/CI
When `ChannelCrashTests` was added, it was never configured to run on Linux/CI. This PR:

- Renames the test case to `ChannelConnectivityTests`
- Adds an entry to `LinuxMain` so this is run on CI
- Makes the class `final`
- Updates indentation to be consistent with the rest of the project (2 spaces)
